+++ ckanext/gobar_theme/helpers/__init__.py
@@ -1,2 +1,3 @@
 # pylint: disable=wildcard-import
 from .helpers import *
+from .organizations import *
